Remove unnecessary or broken code
